[tmux] Switch split pane bindings to open new pane at pwd
As this is what I want to do most of the time, the split pane shortcuts
will open the newly splitted pane at the directory of the pane in which
the shortcuts where issued.

If for whatever reason the default behavior is wanted, there are now
C-b M-v (vertical) and C-b M-h (horizontal) shortcuts.