[qutebrowser] Add userscript + alias for uploading website to remarkable cloud
There is now a userscript `send-to-remarkable` which takes an url and a
tile, then transforms the html from the url to an pdf and sends it to
the remarkable cloud with `rmapi`.

A new alias has also been created. Simply use `send-to-remarkable` to
send the current open tab to the remarkable cloud.