[qutebrowser] Switch search engine prefix to .
To prevent interference between the searx bang search feature and the
way the custom search engines are defined in qutebrowser, the prefix has
been changed to . (dot).

Instead of writing `!clojure defn` to query the clojure docs write
`.clojure defn`.

The dot character should normaly not used in search queries at the first
positions, so it seems perfect for this usecase.