[nvim] Add syntax sync from start to prevent highlighting issues
When scrolling too fast or jumping to a line in a large file, there are
some issues with syntax highlighting. To prevent these, the whole file
is always highlighting from start.

This could lead to some performance issues on large files, but most of
the time it is more important to have consistant syntax highlighting.