[nvim] Fix system clipboard not working sometimes
There was this strange issue that the system clipboard wasn't working
from time to time. In order to get rid of this annoyance, `xsel` is now
configured to be used as system clipboard.

I'm not sure if this is caused by neovim itself or has something to do
with my setup.