[fish] Add wrapper script for ssh to set `TERM`
Alacritty uses its own `TERM` value. This causes some issues with remote
ssh hosts. To prevent those issues there is now a wrapper script for ssh
which sets the `TERM` variable to `xterm-256color`. This has been done
in order to keep the original `TERM` for all local usages, as locally
the right terminfo is installed.