[nvim] Upgrade denite to version containing grep fix
There was an bug with grep for denite which didn't allow it to search
for texts containing whitespaces/special chars in interactive mode and
when used with ripgrep.

This has been fixed with the latest commit on master.

As soon as this commit is part of a new release version, this commit
should be reverted and `tag: '*'` should be used again instead of a
specified commit.